Reader is context-aware, enhanced tests
The main change of this PR is in ReaderFactory:35-40 and addresses the regression.

During the refactoring, ReaderFactory behavior was optimized to
create at most one reader for each type. However, the reader was designed to be context-aware,
otherwise alternatives building does not work properly.

This PR also contains Integration suite enhancements:
 - Symlink current project dir with composer-config-plugin instead of installing it from Packagist
 - Fixed path detection to rely on Composer instead of blindly-guessing it from dirs hierarchy (fixes for previous point)
 - Reduce Integration Test Suite scope to prevent scanning of Integration/Environment/vendor dir